requirements.txt
# COPPERBELT-FRA-SYSTEM Requirements
# Python version: 3.8+
# Last updated: April 2026

# Core Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database & ORM
psycopg2-binary>=2.9.0
django-extensions>=3.2.0

# Task Queue
celery>=5.3.0
redis>=4.5.0

# API & Authentication
dj-rest-auth
django-filter>=23.1

# Data Processing & ML
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dateutil>=2.8.0

# Development & Testing (optional)
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0
